requirements.txt
beautifulsoup4==4.12.3
lxml==5.2.2
pymongo==4.8.0
pytest==8.2.2
python-dotenv==1.0.1
requests==2.32.3